Rename 'person' package to 'client'
