+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -9,7 +9,7 @@ import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
-import seedu.address.model.person.Person;
+import seedu.address.model.client.Person;
 
 /**
  * Deletes a person identified using it's displayed index from the address book.
