+++ src/main/java/seedu/address/model/AddressBook.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 import javafx.collections.ObservableList;
 import seedu.address.commons.util.ToStringBuilder;
-import seedu.address.model.person.Person;
-import seedu.address.model.person.UniquePersonList;
+import seedu.address.model.client.Person;
+import seedu.address.model.client.UniquePersonList;
 
 /**
  * Wraps all data at the address-book level
