+++ src/test/java/seedu/address/logic/commands/AddCommandIntegrationTest.java
@@ -11,7 +11,7 @@ import seedu.address.logic.Messages;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.person.Person;
+import seedu.address.model.client.Person;
 import seedu.address.testutil.PersonBuilder;
 
 /**
