+++ src/test/java/seedu/address/model/person/EmailTest.java
@@ -6,6 +6,8 @@ import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.model.client.Email;
+
 public class EmailTest {
 
     @Test
