+++ src/test/java/seedu/address/model/person/PersonTest.java
@@ -14,6 +14,7 @@ import static seedu.address.testutil.TypicalPersons.BOB;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.model.client.Person;
 import seedu.address.testutil.PersonBuilder;
 
 public class PersonTest {
