+++ src/test/java/seedu/address/testutil/PersonUtil.java
@@ -10,7 +10,7 @@ import java.util.Set;
 
 import seedu.address.logic.commands.AddCommand;
 import seedu.address.logic.commands.EditCommand.EditPersonDescriptor;
-import seedu.address.model.person.Person;
+import seedu.address.model.client.Person;
 import seedu.address.model.tag.Tag;
 
 /**
